% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.misha}
\alias{.misha}
\title{An environment for storing the package global variables}
\format{
An object of class \code{environment} of length 8.
}
\usage{
.misha
}
\description{
An environment for storing the package global variables
}
\keyword{internal}
