% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{gcanvas.load}
\alias{gcanvas.load}
\title{Load a gcanvas.model from disk}
\usage{
gcanvas.load(file)
}
\arguments{
\item{file}{Path to the saved model file}
}
\value{
A gcanvas.model object
}
\description{
Loads a previously saved Markov model from an RDS file.
}
\seealso{
\code{\link{gcanvas.save}}, \code{\link{gcanvas.train}}
}
