% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{gcanvas.save}
\alias{gcanvas.save}
\title{Save a gcanvas.model to disk}
\usage{
gcanvas.save(model, file)
}
\arguments{
\item{model}{A gcanvas.model object from \code{\link{gcanvas.train}}}

\item{file}{Path to save the model}
}
\description{
Saves a trained Markov model to an RDS file for later use.
}
\seealso{
\code{\link{gcanvas.load}}, \code{\link{gcanvas.train}}
}
