% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-root.R
\name{gdb.ls}
\alias{gdb.ls}
\title{List connected databases}
\usage{
gdb.ls()
}
\value{
Character vector of database paths in connection order.
}
\description{
Returns a list of all database paths currently connected to the session.
}
\details{
When multiple databases are connected using \code{gsetroot(c(db1, db2, ...))},
this function returns all database paths in connection order. The first
database is used for shared resources (seq, chrom_sizes), and later databases
take precedence for track resolution ("last wins").
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gdb.ls()
}
\seealso{
\code{\link{gsetroot}}, \code{\link{gdb.info}}, \code{\link{gtrack.db}}
}
\keyword{~database}
\keyword{~db}
