% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-root.R
\name{gdb.summary}
\alias{gdb.summary}
\title{Get summary of connected databases}
\usage{
gdb.summary()
}
\value{
A data frame with columns:
\describe{
  \item{db}{Character: The database path}
  \item{tracks}{Integer: Number of tracks from this database}
  \item{intervals}{Integer: Number of interval sets from this database}
  \item{writable}{Logical: Whether the database's tracks directory is writable}
}
}
\description{
Returns summary information about all connected databases.
}
\details{
This function returns a data frame with information about each connected
database, including the number of tracks, number of interval sets, and
whether the database is writable. Use this when multiple databases are
connected to get an overview of track distribution.

Note: This is different from \code{\link{gdb.info}} which provides detailed
information about a single database's format and structure.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gdb.summary()

}
\seealso{
\code{\link{gdb.ls}}, \code{\link{gsetroot}}, \code{\link{gtrack.db}}, \code{\link{gdb.info}}
}
\keyword{~database}
\keyword{~db}
\keyword{~info}
