% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gdir.create}
\alias{gdir.create}
\title{Creates a new directory in Genomic Database}
\usage{
gdir.create(dir = NULL, showWarnings = TRUE, mode = "0777")
}
\arguments{
\item{dir}{directory path}

\item{showWarnings}{see 'dir.create'}

\item{mode}{see 'dir.create'}
}
\value{
None.
}
\description{
Creates a new directory in Genomic Database.
}
\details{
This function creates a new directory in Genomic Database. Creates only the
last element in the specified path.
}
\note{
A new directory cannot be created within an existing track directory.
}
\seealso{
\code{\link{dir.create}}, \code{\link{gdb.init}},
\code{\link{gdir.cwd}}, \code{\link{gdir.rm}}
}
\keyword{~create}
\keyword{~data}
\keyword{~database}
\keyword{~db}
\keyword{~dir}
\keyword{~directory}
\keyword{~folder}
