% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gdir.cwd}
\alias{gdir.cwd}
\title{Returns the current working directory in Genomic Database}
\usage{
gdir.cwd()
}
\value{
A character string of the path.
}
\description{
Returns the absolute path of the current working directory in Genomic
Database.
}
\details{
This function returns the absolute path of the current working directory in
Genomic Database (not to be confused with shell's current working
directory).
}
\seealso{
\code{\link{gdb.init}}, \code{\link{gdir.cd}},
\code{\link{gdir.create}}, \code{\link{gdir.rm}}
}
\keyword{~cwd}
\keyword{~data}
\keyword{~database}
\keyword{~db}
\keyword{~dir}
\keyword{~directory}
\keyword{~folder}
\keyword{~pwd}
