% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gdir.rm}
\alias{gdir.rm}
\title{Deletes a directory from Genomic Database}
\usage{
gdir.rm(dir = NULL, recursive = FALSE, force = FALSE)
}
\arguments{
\item{dir}{directory path}

\item{recursive}{if 'TRUE', the directory is deleted recursively}

\item{force}{if 'TRUE', suppresses user confirmation of tracks/intervals
removal}
}
\value{
None.
}
\description{
Deletes a directory from Genomic Database.
}
\details{
This function deletes a directory from Genomic Database. If 'recursive' is
'TRUE', the directory is deleted with all the files/directories it contains.
If the directory contains tracks or intervals, the user is prompted to
confirm the deletion. Set 'force' to 'TRUE' to suppress the prompt.
}
\seealso{
\code{\link{gdb.init}}, \code{\link{gdir.create}},
\code{\link{gdir.cd}}, \code{\link{gdir.cwd}}
}
\keyword{~data}
\keyword{~database}
\keyword{~db}
\keyword{~dir}
\keyword{~directory}
\keyword{~folder}
\keyword{~rm}
