\name{gintervals.2d.all}
\alias{gintervals.2d.all}
\title{
Returns 2D intervals that cover the whole genome
}
\description{
Returns 2D intervals that cover the whole genome.
}
\usage{
gintervals.2d.all()
}
\details{
This function returns a set of two-dimensional intervals that cover the whole genome as it is defined by 'chrom_sizes.txt' file.
}
\value{
A data frame representing the intervals.
}

\seealso{
\code{\link{gintervals.2d}}
}
\keyword{ ~genome }
\keyword{ ~chromosome }
\keyword{ ~chromosomes }
\keyword{ ~ALLGENOME }
