% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals-attr.R
\name{gintervals.attr.export}
\alias{gintervals.attr.export}
\title{Returns interval set attributes values}
\usage{
gintervals.attr.export(intervals.set = NULL, attrs = NULL)
}
\arguments{
\item{intervals.set}{a vector of interval set names or 'NULL'}

\item{attrs}{a vector of attribute names or 'NULL'}
}
\value{
A data frame containing interval set attribute values.
}
\description{
Returns interval set attributes values.
}
\details{
This function returns a data frame that contains interval set attribute
values. Column names of the data frame consist of the attribute names, row
names contain the interval set names.

The list of required interval sets is specified by 'intervals.set'
argument. If 'intervals.set' is 'NULL' the attribute values of all existing
interval sets are returned.

Likewise the list of required attributes is controlled by 'attrs' argument.
If 'attrs' is 'NULL' all attribute values of the specified interval sets are
returned. The columns are sorted then by "popularity" of an attribute, i.e.
the number of interval sets containing this attribute. This sorting is not
applied if 'attrs' is not 'NULL'.

Empty character string in a table cell marks a non-existing attribute.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gintervals.attr.set("annotations", "test_attr", "value1")
gintervals.attr.export()
gintervals.attr.export(intervals.set = "annotations")
gintervals.attr.export(attrs = "test_attr")
gintervals.attr.set("annotations", "test_attr", "")

}
\seealso{
\code{\link{gintervals.attr.import}},
\code{\link{gintervals.attr.get}}, \code{\link{gintervals.attr.set}}
}
\keyword{~attr}
\keyword{~attribute}
\keyword{~intervals}
