% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals-attr.R
\name{gintervals.attr.get}
\alias{gintervals.attr.get}
\title{Returns value of an interval set attribute}
\usage{
gintervals.attr.get(intervals.set = NULL, attr = NULL)
}
\arguments{
\item{intervals.set}{interval set name}

\item{attr}{attribute name}
}
\value{
Interval set attribute value (character string).
}
\description{
Returns value of an interval set attribute.
}
\details{
This function returns the value of an interval set attribute. If the
attribute does not exist an empty string is returned.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gintervals.attr.set("annotations", "test_attr", "value")
gintervals.attr.get("annotations", "test_attr")
gintervals.attr.set("annotations", "test_attr", "")

}
\seealso{
\code{\link{gintervals.attr.set}},
\code{\link{gintervals.attr.export}}, \code{\link{gintervals.attr.import}}
}
\keyword{~attr}
\keyword{~attribute}
\keyword{~intervals}
