% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals-attr.R
\name{gintervals.attr.import}
\alias{gintervals.attr.import}
\title{Imports interval set attributes values}
\usage{
gintervals.attr.import(table = NULL, remove.others = FALSE)
}
\arguments{
\item{table}{a data frame containing attribute values}

\item{remove.others}{specifies what to do with the attributes that are not
in the table}
}
\value{
None.
}
\description{
Imports interval set attributes values.
}
\details{
This function imports attribute values contained in a data frame 'table'.
The format of a table is similar to the one returned by
'gintervals.attr.export'. The values of the table must be character strings.
Column names of the table should specify the attribute names, while row
names should contain the interval set names.

The specified attributes of the specified interval sets are modified. If an
attribute value is an empty string this attribute is removed from the
interval set.

If 'remove.others' is 'TRUE' all attributes that do not appear in the table
are removed, otherwise they are preserved unchanged.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
t <- data.frame(
    myattr = "val1", row.names = "annotations1",
    stringsAsFactors = FALSE
)
gintervals.attr.import(t)
gintervals.attr.export(attrs = "myattr")

# roll-back the changes
t$myattr <- ""
gintervals.attr.import(t)

}
\seealso{
\code{\link{gintervals.attr.export}},
\code{\link{gintervals.attr.set}}, \code{\link{gintervals.attr.get}}
}
\keyword{~attr}
\keyword{~attribute}
\keyword{~intervals}
