% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals-attr.R
\name{gintervals.attr.set}
\alias{gintervals.attr.set}
\title{Assigns value to an interval set attribute}
\usage{
gintervals.attr.set(intervals.set = NULL, attr = NULL, value = NULL)
}
\arguments{
\item{intervals.set}{interval set name}

\item{attr}{attribute name}

\item{value}{value (character string)}
}
\value{
None.
}
\description{
Assigns value to an interval set attribute.
}
\details{
This function creates an interval set attribute and assigns 'value' to it.
If the attribute already exists its value is overwritten.

If 'value' is an empty string the attribute is removed.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gintervals.attr.set("annotations", "test_attr", "value")
gintervals.attr.get("annotations", "test_attr")
gintervals.attr.set("annotations", "test_attr", "")

}
\seealso{
\code{\link{gintervals.attr.get}},
\code{\link{gintervals.attr.export}}, \code{\link{gintervals.attr.import}}
}
\keyword{~attr}
\keyword{~attribute}
\keyword{~intervals}
