% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{gintervals.exists}
\alias{gintervals.exists}
\title{Tests for a named intervals set existence}
\usage{
gintervals.exists(intervals.set = NULL)
}
\arguments{
\item{intervals.set}{name of an intervals set}
}
\value{
'TRUE' if a named intervals set exists. Otherwise 'FALSE'.
}
\description{
Tests for a named intervals set existence.
}
\details{
This function returns 'TRUE' if a named intervals set exists in Genomic
Database.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gintervals.exists("annotations")

}
\seealso{
\code{\link{gintervals.ls}}, \code{\link{gintervals.load}},
\code{\link{gintervals.rm}}, \code{\link{gintervals.save}},
\code{\link{gintervals}}, \code{\link{gintervals.2d}}
}
\keyword{~intervals}
