% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{gintervals.is.bigset}
\alias{gintervals.is.bigset}
\title{Tests for big intervals set}
\usage{
gintervals.is.bigset(intervals.set = NULL)
}
\arguments{
\item{intervals.set}{name of an intervals set}
}
\value{
'TRUE' if intervals set is big, otherwise 'FALSE'.
}
\description{
Tests for big intervals set.
}
\details{
This function tests whether 'intervals.set' is a big intervals set.
Intervals set is big if it is stored in big intervals set format and given
the current limits it cannot be fully loaded into memory.

Memory limit is controlled by 'gmax.data.size' option (see:
'getOption("gmax.data.size")').
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gintervals.is.bigset("annotations")

}
\seealso{
\code{\link{gintervals.load}}, \code{\link{gintervals.save}},
\code{\link{gintervals.exists}}, \code{\link{gintervals.ls}}
}
\keyword{~intervals}
