% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{gintervals.save}
\alias{gintervals.save}
\title{Creates a named intervals set}
\usage{
gintervals.save(intervals.set.out = NULL, intervals = NULL)
}
\arguments{
\item{intervals}{intervals}

\item{intervals.set}{name of an intervals set}
}
\value{
None.
}
\description{
Saves intervals to a named intervals set.
}
\details{
This function saves 'intervals' as a named intervals set.
}
\examples{

gdb.init_examples()
intervs <- gintervals(c(1, 2))
gintervals.save("testintervs", intervs)
gintervals.ls()
gintervals.rm("testintervs", force = TRUE)

}
\seealso{
\code{\link{gintervals.rm}}, \code{\link{gintervals.load}},
\code{\link{gintervals.exists}}, \code{\link{gintervals.ls}},
\code{\link{gintervals}}, \code{\link{gintervals.2d}}
}
\keyword{~intervals}
