% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-pred.R
\name{glm_pred.ls}
\alias{glm_pred.ls}
\title{List GLM predictor virtual tracks}
\usage{
glm_pred.ls()
}
\value{
Character vector of virtual track names that use glm.predict.
}
\description{
List GLM predictor virtual tracks
}
