% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-pred.R
\name{glm_pred.rm}
\alias{glm_pred.rm}
\title{Remove a GLM predictor virtual track}
\usage{
glm_pred.rm(name)
}
\arguments{
\item{name}{character(1) Virtual track name}
}
\value{
None.
}
\description{
Remove a GLM predictor virtual track
}
