% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-attr.R
\name{gtrack.attr.get}
\alias{gtrack.attr.get}
\title{Returns value of a track attribute}
\usage{
gtrack.attr.get(track = NULL, attr = NULL)
}
\arguments{
\item{track}{track name}

\item{attr}{attribute name}
}
\value{
Track attribute value.
}
\description{
Returns value of a track attribute.
}
\details{
This function returns the value of a track attribute. If the attribute does
not exist an empty sting is returned.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gtrack.attr.set("sparse_track", "test_attr", "value")
gtrack.attr.get("sparse_track", "test_attr")
gtrack.attr.set("sparse_track", "test_attr", "")

}
\seealso{
\code{\link{gtrack.attr.import}}, \code{\link{gtrack.attr.set}}
}
\keyword{~attr}
\keyword{~attribute}
