% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gtrack.create_dirs}
\alias{gtrack.create_dirs}
\title{Create directories needed for track creation}
\usage{
gtrack.create_dirs(track, mode = "0777")
}
\arguments{
\item{track}{name of the track}

\item{mode}{see 'dir.create'}
}
\value{
None.
}
\description{
This function creates the directories needed for track creation.
For example, if the track name is 'proj.sample.my_track', this function
creates the directories 'proj' and 'sample'. Use this function with caution -
a long track name may create a deep directory structure.
}
\examples{

gdb.init_examples()

# This creates the directories 'proj' and 'sample'
gtrack.create_dirs("proj.sample.my_track")

}
