% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-management.R
\name{gtrack.mv}
\alias{gtrack.mv}
\title{Renames or moves a track}
\usage{
gtrack.mv(src = NULL, dest = NULL)
}
\arguments{
\item{src}{source track name, optionally with prefix (e.g., "at@my.track")}

\item{dest}{destination track name, optionally with prefix (e.g., "al@my.track")}
}
\value{
None.
}
\description{
Renames a track or moves it to a different namespace within the same database,
or moves it between databases when using prefixed names.
}
\details{
This function renames a track or moves it to a different namespace (directory).
When both source and destination have the same prefix (or both are unprefixed),
a simple file rename is performed. When different prefixes are used
(e.g., "src@track" to "dst@track"), the track is copied to the destination
database and then deleted from the source.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gtrack.create_sparse("test_track", "Test", gintervals(1, 0, 100), 1)
gtrack.mv("test_track", "renamed_track")
gtrack.exists("renamed_track")
gtrack.rm("renamed_track", force = TRUE)

}
\seealso{
\code{\link{gtrack.copy}}, \code{\link{gtrack.rm}},
\code{\link{gtrack.exists}}, \code{\link{gtrack.ls}}
}
\keyword{~track}
