% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrack.R
\name{gvtrack.clear}
\alias{gvtrack.clear}
\title{Deletes all virtual tracks}
\usage{
gvtrack.clear()
}
\value{
None.
}
\description{
Deletes all virtual tracks of the current working directory.
}
\details{
This function removes every virtual track defined for the current working
directory in one call. After it returns \code{\link{gvtrack.ls}} reports no
virtual tracks. It is a convenient way to reset virtual-track state between
analyses. Use \code{\link{gvtrack.rm}} to remove a single virtual track.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gvtrack.create("vtrack1", "dense_track", "max")
gvtrack.create("vtrack2", "dense_track", "avg")
gvtrack.ls()
gvtrack.clear()
gvtrack.ls()

}
\seealso{
\code{\link{gvtrack.rm}}, \code{\link{gvtrack.ls}},
\code{\link{gvtrack.create}}
}
\keyword{~virtual}
