% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{print.gcanvas.model}
\alias{print.gcanvas.model}
\title{Print summary of a gcanvas.model}
\usage{
\method{print}{gcanvas.model}(x, ...)
}
\arguments{
\item{x}{A gcanvas.model object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print summary of a gcanvas.model
}
