% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{print.gsynth.model}
\alias{print.gsynth.model}
\title{Print summary of a gsynth.model}
\usage{
\method{print}{gsynth.model}(x, ...)
}
\arguments{
\item{x}{A gsynth.model object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print summary of a gsynth.model
}
