% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{gdataset.example_path}
\alias{gdataset.example_path}
\title{Create an example dataset on the fly}
\usage{
gdataset.example_path()
}
\value{
Path to the created dataset directory
}
\description{
Creates a small dataset in a temporary directory using the built-in
example database. The working database is reset via
\code{\link{gdb.init_examples}}.
Temporary tracks and intervals are removed after saving to avoid
name collisions.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

dataset_path <- gdataset.example_path()
gdataset.load(dataset_path)
gdataset.unload(dataset_path)

}
\seealso{
\code{\link{gdataset.save}}, \code{\link{gdataset.load}},
  \code{\link{gdb.init_examples}}
}
