% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{gdataset.info}
\alias{gdataset.info}
\title{Get dataset information}
\usage{
gdataset.info(path)
}
\arguments{
\item{path}{Path to any dataset (loaded or not)}
}
\value{
List with dataset information
}
\description{
Returns metadata and contents of a dataset.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

dataset_path <- gdataset.example_path()
gdataset.info(dataset_path)

}
\seealso{
\code{\link{gdataset.ls}}, \code{\link{gdataset.load}}
}
