% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{gdataset.ls}
\alias{gdataset.ls}
\title{List working database and loaded datasets}
\usage{
gdataset.ls(dataframe = FALSE)
}
\arguments{
\item{dataframe}{If FALSE, return character vector; if TRUE, return data frame}
}
\value{
Character vector of paths or data frame with detailed information
}
\description{
Returns a list of the working database and all loaded datasets.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

dataset_path <- gdataset.example_path()
gdataset.load(dataset_path)
gdataset.ls()
gdataset.unload(dataset_path)

}
\seealso{
\code{\link{gdataset.load}}, \code{\link{gdataset.info}}
}
