% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gdb.get_readonly_attrs}
\alias{gdb.get_readonly_attrs}
\title{Returns a list of read-only track attributes}
\usage{
gdb.get_readonly_attrs()
}
\value{
A list of read-only track attributes.
}
\description{
Returns a list of read-only track attributes.
}
\details{
This function returns a list of read-only track attributes. These attributes
are not allowed to be modified or deleted.

If no attributes are marked as read-only a 'NULL' is returned.
}
\seealso{
\code{\link{gdb.set_readonly_attrs}},
\code{\link{gtrack.attr.get}}, \code{\link{gtrack.attr.set}}
}
\keyword{~attr}
\keyword{~attribute}
