% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gdb.set_readonly_attrs}
\alias{gdb.set_readonly_attrs}
\title{Sets read-only track attributes}
\usage{
gdb.set_readonly_attrs(attrs)
}
\arguments{
\item{attrs}{a vector of read-only attributes names or 'NULL'}
}
\value{
None.
}
\description{
Sets read-only track attributes.
}
\details{
This function sets the list of read-only track attributes. The specified
attributes may or may not already exist in the tracks.

If 'attrs' is 'NULL' the list of read-only attributes is emptied.
}
\seealso{
\code{\link{gdb.get_readonly_attrs}},
\code{\link{gtrack.attr.get}}, \code{\link{gtrack.attr.set}}
}
\keyword{~attr}
\keyword{~attribute}
