\name{gdir.cwd}
\alias{gdir.cwd}
\title{
Returns the current working directory in Genomic Database
}
\description{
Returns the absolute path of the current working directory in Genomic Database.
}
\usage{
gdir.cwd()
}
\details{
This function returns the absolute path of the current working directory in Genomic Database (not to be confused with shell's current working directory).
}
\value{
A character string of the path.
}

\seealso{
\code{\link{gdb.init}}, \code{\link{gdir.cd}}, \code{\link{gdir.create}}, \code{\link{gdir.rm}}
}
\keyword{ ~db }
\keyword{ ~data }
\keyword{ ~database }
\keyword{ ~cwd }
\keyword{ ~pwd }
\keyword{ ~dir }
\keyword{ ~directory }
\keyword{ ~folder }
