% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{gintervals.2d.all}
\alias{gintervals.2d.all}
\title{Returns 2D intervals that cover the whole genome}
\usage{
gintervals.2d.all()
}
\value{
A data frame representing the intervals.
}
\description{
Returns 2D intervals that cover the whole genome.
}
\details{
This function returns a set of two-dimensional intervals that cover the
whole genome as it is defined by 'chrom_sizes.txt' file.
}
\seealso{
\code{\link{gintervals.2d}}
}
\keyword{~ALLGENOME}
\keyword{~chromosome}
\keyword{~chromosomes}
\keyword{~genome}
