\name{gintervals.all}
\alias{gintervals.all}
\title{
Returns 1D intervals that cover the whole genome
}
\description{
Returns 1D intervals that cover the whole genome.
}
\usage{
gintervals.all()
}
\details{
This function returns a set of one-dimensional intervals that cover the whole genome as it is defined by 'chrom_sizes.txt' file.
}
\value{
A data frame representing the intervals.
}

\seealso{
\code{\link{gintervals}}
}
\keyword{ ~genome }
\keyword{ ~chromosome }
\keyword{ ~chromosomes }
\keyword{ ~ALLGENOME }
