% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{gintervals.chrom_sizes}
\alias{gintervals.chrom_sizes}
\title{Returns number of intervals per chromosome}
\usage{
gintervals.chrom_sizes(intervals = NULL)
}
\arguments{
\item{intervals}{intervals set}
}
\value{
Data frame representing number of intervals per chromosome (for 1D
intervals) or chromosome pair (for 2D intervals).
}
\description{
Returns number of intervals per chromosome (or chromosome pair).
}
\details{
This function returns number of intervals per chromosome (for 1D intervals)
or chromosome pair (for 2D intervals).
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gintervals.chrom_sizes("annotations")

}
\seealso{
\code{\link{gintervals.load}}, \code{\link{gintervals.save}},
\code{\link{gintervals.exists}}, \code{\link{gintervals.ls}},
\code{\link{gintervals}}, \code{\link{gintervals.2d}}
}
\keyword{~intervals}
