% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals-management.R
\name{gintervals.save}
\alias{gintervals.save}
\title{Creates a named intervals set}
\usage{
gintervals.save(intervals.set.out = NULL, intervals = NULL)
}
\arguments{
\item{intervals.set.out}{name of the new intervals set, optionally with prefix
(e.g., "at@my.intervals" to save in database with prefix "at")}

\item{intervals}{intervals to save}
}
\value{
None.
}
\description{
Saves intervals to a named intervals set.
}
\details{
This function saves 'intervals' as a named intervals set. The destination
database is determined by the prefix in the intervals set name. If the name
has a prefix (e.g., "at@my.intervals"), the intervals are saved in that database.
If no prefix is specified, the intervals are saved in the current working directory.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
intervs <- gintervals(c(1, 2))
gintervals.save("testintervs", intervs)
gintervals.ls()
gintervals.rm("testintervs", force = TRUE)

}
\seealso{
\code{\link{gintervals.rm}}, \code{\link{gintervals.load}},
\code{\link{gintervals.exists}}, \code{\link{gintervals.ls}},
\code{\link{gintervals}}, \code{\link{gintervals.2d}}
}
\keyword{~intervals}
