% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{gsynth.load}
\alias{gsynth.load}
\title{Load a gsynth.model from disk}
\usage{
gsynth.load(file)
}
\arguments{
\item{file}{Path to the saved model file}
}
\value{
A gsynth.model object
}
\description{
Loads a previously saved Markov model from an RDS file.
}
\seealso{
\code{\link{gsynth.save}}, \code{\link{gsynth.train}}
}
