% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{gsynth.save}
\alias{gsynth.save}
\title{Save a gsynth.model to disk}
\usage{
gsynth.save(model, file)
}
\arguments{
\item{model}{A gsynth.model object from \code{\link{gsynth.train}}}

\item{file}{Path to save the model}
}
\description{
Saves a trained Markov model to an RDS file for later use.
}
\seealso{
\code{\link{gsynth.load}}, \code{\link{gsynth.train}}
}
