% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gtraceback}
\alias{gtraceback}
\title{Prints call stack of the last uncaught error}
\usage{
gtraceback(x = NULL, max.lines = getOption("deparse.max.lines"))
}
\arguments{
\item{x}{see 'traceback'}

\item{max.lines}{see 'traceback'}
}
\value{
See 'traceback'.
}
\description{
Prints call stack of the last uncaught error in a friendly way.
}
\details{
Similarly to 'traceback' this function prints the call stack of the last
uncaught error. Yet 'gtraceback' does it in a more friendly way by omitting
the calls that occurred inside the library.
}
\examples{

gdb.init_examples()
f <- function() {
    gscreen("blablabla")
}
f()
gtraceback()

}
\seealso{
\code{\link{traceback}}
}
\keyword{~error}
\keyword{~exception}
\keyword{~trace}
