% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-array.R
\name{gtrack.array.get_colnames}
\alias{gtrack.array.get_colnames}
\title{Returns column names of array track}
\usage{
gtrack.array.get_colnames(track = NULL)
}
\arguments{
\item{track}{track name}
}
\value{
A character vector with column names.
}
\description{
Returns column names of array track.
}
\details{
This function returns the column names of an array track.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gtrack.array.get_colnames("array_track")

}
\seealso{
\code{\link{gtrack.array.set_colnames}},
\code{\link{gtrack.array.extract}}, \code{\link{gvtrack.array.slice}},
\code{\link{gtrack.info}}
}
\keyword{~array}
\keyword{~columns}
