% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{gtrack.convert}
\alias{gtrack.convert}
\title{Converts a track to the most current format}
\usage{
gtrack.convert(src.track = NULL, tgt.track = NULL)
}
\arguments{
\item{src.track}{source track name}

\item{tgt.track}{target track name. If 'NULL' the source track is
overwritten.}
}
\value{
None
}
\description{
Converts a track (if needed) to the most current format.
}
\details{
This function converts a track to the most current format. It should be used
if a track created by an old version of the library cannot be read anymore
by the newer version. The old track is given by 'src.track'. After
conversion a new track 'tgt.track' is created. If 'tgt.track' is 'NULL' the
source track is overwritten.
}
\seealso{
\code{\link{gtrack.create}}, \code{\link{gtrack.2d.create}},
\code{\link{gtrack.create_sparse}}
}
\keyword{~convert}
\keyword{~track}
