% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-management.R
\name{gtrack.copy}
\alias{gtrack.copy}
\title{Copies a track}
\usage{
gtrack.copy(src = NULL, dest = NULL)
}
\arguments{
\item{src}{source track name}

\item{dest}{destination track name}
}
\value{
None.
}
\description{
Creates a copy of an existing track.
}
\details{
This function creates a copy of a track. The new track is created in the
current working directory (.misha$GWD), which may be in a different database than
the source track when multiple databases are connected.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gtrack.copy("dense_track", "dense_track_copy")
gtrack.exists("dense_track_copy")
gtrack.rm("dense_track_copy", force = TRUE)

}
\seealso{
\code{\link{gtrack.mv}}, \code{\link{gtrack.rm}},
\code{\link{gtrack.exists}}, \code{\link{gtrack.ls}}
}
\keyword{~track}
