% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-management.R
\name{gtrack.copy}
\alias{gtrack.copy}
\title{Copies a track}
\usage{
gtrack.copy(src = NULL, dest = NULL)
}
\arguments{
\item{src}{source track name, optionally with prefix (e.g., "at@my.track")}

\item{dest}{destination track name, optionally with prefix (e.g., "al@my.track")}
}
\value{
None.
}
\description{
Creates a copy of an existing track.
}
\details{
This function creates a copy of a track. The destination database is determined
by the prefix in the destination name. If the destination has a prefix
(e.g., "al@track"), the track is created in that database. If no prefix is
specified, the track is created in the current working directory (.misha$GWD).

This allows copying tracks between databases:
\code{gtrack.copy("src@track", "dst@track")} copies from the "src" database
to the "dst" database.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gtrack.copy("dense_track", "dense_track_copy")
gtrack.exists("dense_track_copy")
gtrack.rm("dense_track_copy", force = TRUE)

}
\seealso{
\code{\link{gtrack.mv}}, \code{\link{gtrack.rm}},
\code{\link{gtrack.exists}}, \code{\link{gtrack.ls}}
}
\keyword{~track}
