% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{gtrack.exists}
\alias{gtrack.exists}
\title{Tests for a track existence}
\usage{
gtrack.exists(track = NULL)
}
\arguments{
\item{track}{track name}
}
\value{
'TRUE' if a track exists. Otherwise 'FALSE'.
}
\description{
Tests for a track existence.
}
\details{
This function returns 'TRUE' if a track exists in Genomic Database.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gtrack.exists("dense_track")

}
\seealso{
\code{\link{gtrack.ls}}, \code{\link{gtrack.info}},
\code{\link{gtrack.create}}, \code{\link{gtrack.rm}}
}
\keyword{~track}
