% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-management.R
\name{gtrack.ls}
\alias{gtrack.ls}
\title{Returns a list of track names}
\usage{
gtrack.ls(
  ...,
  pattern = NULL,
  db = NULL,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{...}{additional arguments of the form 'attribute = pattern', used to
filter tracks by track attribute value}

\item{pattern}{optional pattern to match against track names, equivalent to
passing it positionally (e.g. \code{gtrack.ls("den*")} and
\code{gtrack.ls(pattern = "den*")} are the same). Note that this means a
track attribute literally named "pattern" can no longer be filtered by name
via \code{...}; use \code{gtrack.ls()} followed by
\code{\link{gtrack.attr.get}} instead.}

\item{db}{optional database path to filter tracks. If specified, only tracks
from that database are returned.}

\item{ignore.case, perl, fixed, useBytes}{see 'grep'}
}
\value{
An array that contains the names of tracks that match the supplied
patterns.
}
\description{
Returns a list of track names in Genomic Database.
}
\details{
This function returns a list of tracks whose name or track attribute value
match a pattern (see 'grep'). If called without any arguments all tracks are
returned.

If pattern is specified without a track attribute (i.e. in the form of
'pattern', either positionally or via the 'pattern' argument) then filtering
is applied to the track names. If pattern is supplied with a track attribute
(i.e. in the form of 'name = pattern') then track attribute is matched
against the pattern.

Multiple patterns are applied one after another. The resulted list of tracks
should match all the patterns.

When multiple databases are connected, the 'db' parameter can be used to
filter tracks to only those from a specific database.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()

# get all track names
gtrack.ls()

# get track names that match the pattern "den*"
gtrack.ls("den*")

# equivalent, using the 'pattern' argument
gtrack.ls(pattern = "den*")

# get track names whose "created.by" attribute match the pattern
# "create_sparse"
gtrack.ls(created.by = "create_sparse")

# get track names whose names match the pattern "den*" and whose
# "created.by" attribute match the pattern "track"
gtrack.ls("den*", created.by = "track")

}
\seealso{
\code{\link{grep}}, \code{\link{gtrack.exists}},
\code{\link{gtrack.create}}, \code{\link{gtrack.rm}}, \code{\link{gtrack.dataset}}
}
\keyword{~intervals}
\keyword{~ls}
