% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-var.R
\name{gtrack.var.rm}
\alias{gtrack.var.rm}
\title{Deletes a track variable}
\usage{
gtrack.var.rm(track = NULL, var = NULL)
}
\arguments{
\item{track}{track name}

\item{var}{track variable name}
}
\value{
None.
}
\description{
Deletes a track variable.
}
\details{
This function deletes a track variable.
}
\examples{
\dontshow{
options(gmax.processes = 2)
}

gdb.init_examples()
gtrack.var.set("sparse_track", "test_var1", 1:10)
gtrack.var.set("sparse_track", "test_var2", "v")
gtrack.var.ls("sparse_track")
gtrack.var.rm("sparse_track", "test_var1")
gtrack.var.rm("sparse_track", "test_var2")
gtrack.var.ls("sparse_track")

}
\seealso{
\code{\link{gtrack.var.get}}, \code{\link{gtrack.var.set}},
\code{\link{gtrack.var.ls}}
}
\keyword{~variable}
