% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misha.R
\name{gvtrack.info}
\alias{gvtrack.info}
\title{Returns the definition of a virtual track}
\usage{
gvtrack.info(vtrack = NULL)
}
\arguments{
\item{vtrack}{virtual track name}
}
\value{
Internal representation of a virtual track.
}
\description{
Returns the definition of a virtual track.
}
\details{
This function returns the internal represenation of a virtual track.
}
\examples{

gdb.init_examples()
gvtrack.create("vtrack1", "dense_track", "max")
gvtrack.info("vtrack1")

}
\seealso{
\code{\link{gvtrack.create}}
}
\keyword{~virtual}
